requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.7.0
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
bidict==0.22.1
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
dataclasses-json==0.5.7
docopt==0.6.2
exceptiongroup==1.1.1
fastapi==0.96.0
frozenlist==1.3.3
gevent==22.10.2
gitdb==4.0.10
GitPython==3.1.31
greenlet==2.0.2
grpcio==1.54.2
grpcio-tools==1.54.2
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.17.2
httpx==0.24.1
hyperframe==6.0.1
idna==3.4
iniconfig==2.0.0
Jinja2==3.1.2
python-jose==3.3.0
langchain==0.0.217
langchainplus-sdk==0.0.17
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
multidict==6.0.4
mypy-extensions==1.0.0
nest-asyncio==1.5.6
numexpr==2.8.4
numpy==1.24.3
openai==0.27.7
openapi-schema-pydantic==1.2.4
packaging==23.1
passlib==1.7.4
pipreqs==0.4.13
pluggy==1.0.0
portalocker==2.7.0
protobuf==4.23.2
pydantic==1.10.8
pytest==7.3.1
python-engineio==4.4.1
python-multipart==0.0.6
python-socketio==5.8.0
python-sonarqube-api==2.0.3
PyYAML==6.0
qdrant-client==1.2.0
regex==2023.5.5
requests==2.30.0
smmap==5.0.0
sniffio==1.3.0
SQLAlchemy==2.0.15
starlette==0.27.0
tenacity==8.2.2
tiktoken==0.4.0
tomli==2.0.1
tqdm==4.65.0
tree-sitter==0.20.1
typing-inspect==0.9.0
typing_extensions==4.5.0
urllib3==1.26.16
uvicorn==0.22.0
websocket==0.2.1
yarg==0.1.9
yarl==1.9.2
zope.event==4.6
zope.interface==6.0
psycopg2==2.9.6
alembic==1.11.1
redis==4.6.0